Have MIDAS define the right constant area start symbol.
When building MUDDLE, I was seeing various unresolved symbols of the
form "$   42". These are used by MIDAS when it wants to refer to
something in a CONSTANTS area, but doesn't know the location at assembly
time (for example, because it's the other side of a LOC X where X is a
global). At the start of a CONSTANTS area, MIDAS should define this
symbol as .; however, it was defining it with the wrong name.

There's a special case in RPWRD1 to output these symbols as squoze
rather than symbol table numbers -- but the code in CNST2 that defines
the symbol was using PDEFPT, which always outputs the number.
Modify it to output the squoze instead.